fix(voi): give a volume the VOI LUT Sequence of every provider
The volume path accepted only the already parsed shape that the wadouri
provider gives. The wadors provider passes the raw DICOMweb JSON items
through, and the dcmjs providers give LUTDescriptor and LUTData. Both of
those shapes failed isRenderableVOILUT. A volume that streams its own
instances also puts no image in the image cache, so the image was not a
source either. Thus a volume of every provider except wadouri lost the VOI
LUT Sequence of the file without a message, and it rendered a linear
window in place of the curve.

The image loader already had a normalizer for all three shapes. This
commit moves that normalizer to the core package, where the volume path
reaches it also, and the volume path now uses it. The module in the image
loader stays as a re-export, so the import path of the loader does not
change.

Two more corrections in BaseVolumeViewport:

- _getVolumeVOIShape kept the first shape that it read. The wadouri
  provider reads the VOI from the file of the instance, and that file
  arrives after the volume. Thus a shape that was read while the volume
  loaded could be empty, and the volume kept that empty answer for the
  life of the viewport. The shape is now read one more time when the load
  completes.

- addVolumes adds an actor with addActors, and it did not record which
  actors setDefaultVolumeVOI gave a curve to. Thus a fusion volume whose
  file gives a sigmoid or a VOI LUT Sequence had no record, and
  getProperties reported the whole node domain of the curve as the VOI -
  about 3.3 times the window width for a sigmoid. The window level tool
  and setVOILUTFunction both start from that value. Both actor paths now
  call one method, _recordVolumeVOICurves.

The VOI_MODIFIED detail of a volume also reported voiLUTSequenceApplied
from the file and not from the actor, so it stayed true when a colormap
replaced the curve. It now reports the applied state, as StackViewport
does.
